Add constraints to models

+++ app/backend/backend/Dtos/Account/ClientProfileDto.cs
@@ -4,7 +4,7 @@ namespace backend.Dtos.Account
     {
         public int OrdersCount { get; set; }
 
-        public float OrdersCost { get; set; }
+        public decimal OrdersCost { get; set; }
 
         public int ProductsInCartCount { get; set; }
     }
