+++ app/backend/backend/Dtos/Order/OrderItemDto.cs
@@ -5,8 +5,8 @@ namespace backend.Dtos.Order
     public class OrderItemDto
     {
         public ProductDto Product { get; set; }
-        public float ProductPrice { get; set; }
+        public decimal ProductPrice { get; set; }
         public int Quantity { get; set; }
-        public float FinalPrice => ProductPrice * Quantity;
+        public decimal FinalPrice => ProductPrice * Quantity;
     }
 }
